+++ samples/snippets/downloadMergedDocumentGroup.js
@@ -4,20 +4,22 @@ const signnow = require('@signnow/api-client')({
   credentials: 'BASE64_ENCODED_CLIENT_CREDENTIALS',
   production: true, // if false then uses eval server
 });
-const downloadDocumentGroup = signnow.documentGroup.download;
+const { download: downloadDocumentGroup } = signnow.documentGroup;
 
 const id = 'DOCUMENT_GROUP_ID_GOES_HERE';
 const token = 'YOUR_ACCESS_TOKEN';
-const document_order = [ // specifys the order in which group's documents will be merged
-  'ID_OF_THE_FIRST_DOCUMENT',
-  'ID_OF_THE_SECOND_DOCUMENT',
+
+// specify the order in which group's documents will be merged
+const document_order = [
+  'DOCUMENT_ID_1',
+  'DOCUMENT_ID_2',
 ];
 
 /**
- * @param {Binary} res - binary data (PDF file) of the document group
+ * @param {Buffer} res - binary data of the merged Document Group (PDF file)
  */
 const handleResponse = res => {
-  console.log(res);
+  // save the file to your disk or pipe it to another handler
 };
 
 const handleError = err => {
@@ -27,8 +29,8 @@ const handleError = err => {
 downloadDocumentGroup({
   id,
   token,
-  type: 'merged', // 'zip' is default
-  with_history: 'after_merged_pdf', // 'no' is default, can also be set to 'after_each_document'
+  type: 'merged',
+  with_history: 'after_merged_pdf',
   document_order,
 }, (err, res) => {
   if (err) {